5.7 Создал класс TripController и перенес весь render связанный с Trip туда.
